refactor: flatten normal dispatch modules
