+++ crates/evm2/src/interpreter/instructions/table/normal_single_return.rs
@@ -5,12 +5,15 @@ use crate::{
 };
 use core::hint::cold_path;
 
-/// Unpacked normal instruction return value.
-type InstrFnRet = (*const u8, usize);
-
-/// Unpacked normal instruction function pointer.
-pub(super) type RawInstrFn<T> =
-    extern_table!(fn(pc: Pc, stack: Stack<'_>, state: &mut InterpreterState<'_, T>) -> InstrFnRet);
+/// Single-return normal instruction function pointer.
+pub(super) type RawInstrFn<T> = extern_table!(
+    fn(
+        pc: Pc,
+        stack: Stack<'_>,
+        state: &mut InterpreterState<'_, T>,
+        next_stack_len: &mut usize,
+    ) -> Pc
+);
 
 extern_table! {
     pub(super) fn dispatch<
@@ -23,9 +26,10 @@ extern_table! {
         pc: Pc,
         stack: Stack<'_>,
         state: &mut InterpreterState<'_, T>,
-    ) -> InstrFnRet {
+        next_stack_len: &mut usize,
+    ) -> Pc {
         let _ = DYNAMIC_GAS;
-        dispatch_mono::<T, C, M>(pc, stack, state, OP)
+        dispatch_mono::<T, C, M>(pc, stack, state, next_stack_len, OP)
     }
 }
 
@@ -35,8 +39,9 @@ fn dispatch_mono<T: EvmTypes, C: EvmConfig<T>, M: InspectMode<T>>(
     mut pc: Pc,
     mut stack: Stack<'_>,
     state: &mut InterpreterState<'_, T>,
+    next_stack_len: &mut usize,
     op: u8,
-) -> InstrFnRet {
+) -> Pc {
     let instr = C::VERSION_TABLES.instruction(op).instr;
     if M::INSPECT {
         M::step(state, pc, stack.len);
@@ -46,7 +51,7 @@ fn dispatch_mono<T: EvmTypes, C: EvmConfig<T>, M: InspectMode<T>>(
         Ok(()) => {
             r = instr(&mut pc, stack.as_mut(), state);
             if !M::INSPECT || r.is_ok() {
-                super::super::inc_pc(&mut pc, op);
+                super::inc_pc(&mut pc, op);
             }
         }
         Err(e) => r = Err(e),
@@ -55,14 +60,15 @@ fn dispatch_mono<T: EvmTypes, C: EvmConfig<T>, M: InspectMode<T>>(
         state.set_result(r);
         M::step_end(state, pc, stack.len);
     }
+    *next_stack_len = stack.len;
     if r.is_err() {
         cold_path();
         if !M::INSPECT {
             state.set_result(r);
         }
-        return (core::ptr::null(), stack.len);
+        return Pc::new(core::ptr::null());
     }
-    (pc.as_ptr(), stack.len)
+    pc
 }
 
 #[inline(always)]
